scripts changes

+++ misc/00-one-time/tmp-huayang-folder-sync.php
@@ -1,168 +1,93 @@
 <?php
 
 $dirs = <<<EOF
-爱蜜社_7Seven《制服丝袜美腿》-Vol.079-写真集-57P-
-爱蜜社_ALICE梁紫轩《空姐制服诱惑》-Vol.013-写真集-56P-
-爱蜜社_ALICE梁紫轩---蕾丝内衣-Vol.003-写真集-61P-
-爱蜜社_Angela小热巴《透明的肚兜诱惑》-Vol.285-写真集-40P-
-爱蜜社_Carry《魅惑内衣-丝袜美足》-Vol.397-写真集-54P-
-爱蜜社_ceci《身材高挑曼妙凹凸别致》-Vol.395-写真集-40P-
-爱蜜社_Evon陈赞之《韵味旗袍与朦胧丝袜诱惑》-Vol.403-写真集-41P-
-爱蜜社_LalaBaby啦啦《两套内衣-护士装-OL装》-Vol.116-写真集-75P-
-爱蜜社_LalaBaby啦啦《傲人胸围-修长美腿》-Vol.044-写真集-48P-
-爱蜜社_LalaBaby啦啦《普吉岛旅拍》清新泳装小诱惑-Vol.051-写真集-54P-
-爱蜜社_LalaBaby啦啦《清凉比基尼-牛仔热裤》-Vol.039-写真集-55P-
-爱蜜社_Lavinia-Cccil-《魔鬼的身材，天使的面孔》-Vol.343-写真集-38P-
-爱蜜社_Lavinia《丝袜美腿特辑》-Vol.384-写真集-36P-
-爱蜜社_Lavinia《古典韵味旗袍-丝袜美腿》-Vol.362-写真集-47P-
-爱蜜社_Lavinia《无法抵御的职业装》-Vol.408-写真集-62P-
-爱蜜社_Lavinia《牛仔长腿-丝袜美腿》-Vol.367-写真集-45P-
-爱蜜社_Lavinia肉肉《朦胧丝袜让人心神俱醉》-Vol.433-写真集-39P-
-爱蜜社_Lavinia肉肉《粉色连体裙包裹下的女神》-Vol.439-写真集-57P-
-爱蜜社_Lavinia《魅惑的丝袜美腿极致销魂》-Vol.376-写真集-41P-
-爱蜜社_luvian本能《两套性感服饰》-Vol.385-写真集-54P-
-爱蜜社_Lynn刘奕宁《一朵含苞待放的牡丹花》-Vol.368-写真集-39P-
-爱蜜社_Michi_Ho《比基尼-OL丝袜装》-Vol.050-写真集-55P-
-爱蜜社_Panadol雅《女仆装-OL装》-Vol.099-写真集-65P-
-爱蜜社_Sandy陈天扬《热裤-性感蕾丝-白衬衫》-Vol.072-写真集-53P-
-爱蜜社_Sandy陈天扬《花仙子般比基尼系列》-Vol.087-写真集-50P-
-爱蜜社_SISY思《典雅旗袍-诱惑丝袜-性感内衣》-Vol.117-写真集-54P-
-爱蜜社_SISY思《外拍2套清新泳装》-Vol.065-写真集-53P-
-爱蜜社_SISY思《旗袍-晚礼裙》-Vol.030-写真集-53P-
-爱蜜社_SISY思《黑丝OL》-Vol.027-写真集-51P-
-爱蜜社_sugar小甜心CC《2套丝袜诱惑系列》-Vol.132-写真集-53P-
-爱蜜社_sugar小甜心CC《丝袜美腿-性感美臀》-Vol.129-写真集-49P-
-爱蜜社_sugar小甜心CC《丝袜美腿-浴室惹火诱惑》-Vol.127-写真集-42P-
-爱蜜社_sugar小甜心CC《室拍2套性感内衣》-Vol.125-写真集-53P-
-爱蜜社_sugar小甜心CC《巴厘岛旅拍》两套性感内衣-Vol.121-写真集-58P-
-爱蜜社_sugar小甜心CC《巴厘岛旅拍》美臀诱惑-Vol.118-写真集-45P-
-爱蜜社_sugar小甜心CC《性感比基尼-美艳厨娘装》-Vol.113-写真集-57P-
-爱蜜社_sugar小甜心CC《性感甜美的小护士》-Vol.093-写真集-50P-
-爱蜜社_sugar小甜心CC《甜美诱人的小甜心》-Vol.120-写真集-63P-
-爱蜜社_sugar小甜心CC《空姐制服-厨娘服》-Vol.073-写真集-61P-
-爱蜜社_sugar小甜心CC《粉色丝袜-性感内衣》-Vol.100-写真集-43P-
-爱蜜社_sugar小甜心CC《美女扮野兽-旗袍与丝袜》-Vol.136-写真集-59P-
-爱蜜社_-Vol.060-于姬、朱若慕、徐妍馨、傅诗瑶、小丽莎萌萌哒及其他模特合集-78P-
-爱蜜社_-Vol.068-刘奕宁、夏茉、其他模特合集-49P-
-爱蜜社_-Vol.143-SISY思、Panadol雅、刘奕宁、LalaBaby啦啦等十二位模特合集-53P-
-爱蜜社_-Vol.465-LindaLinda-《性感吊带肉丝袜美腿》-写真集-69P-
-爱蜜社_-Vol.479-Angela小热巴---灰色风格的都市丝袜美腿写真-64P-
-爱蜜社_-Vol.489-Angela小热巴---丝足美腿灵动曼妙-108P-
-爱蜜社_-Vol.494-Lavinia肉肉---丝袜美腿特辑写真-53P-
-爱蜜社_-Vol.496-evon---校服淋雨私房主题系列-65P-
-爱蜜社_-Vol.498-杨紫嫣Cynthia---典雅礼裙与极致魅惑丝袜-54P-
-爱蜜社_-Vol.500-Evon陈赞之---眼镜OL主题系列-43P-
-爱蜜社_-Vol.501-蜜桃cc---丝袜主题极致魅惑-76P-
-爱蜜社_-Vol.502-杨紫嫣Cynthia---典雅礼裙与极致魅惑丝袜-55P-
-爱蜜社_-Vol.503-九月生_---吊带与极致黑丝的视觉体验-95P-
-爱蜜社_-Vol.504--Lavinia肉肉和-蜜桃cc---酒店私房0L系列-66P-
-爱蜜社_-Vol.505-方子萱---条纹的职场秘书OL系列-73P-
-爱蜜社_-Vol.506-九月生_---白衬衫黑短裙职场秘书制服系列-53P-
-爱蜜社_-Vol.507-Lavinia肉肉---职场OL风格丝袜美腿-54P-
-爱蜜社_-Vol.508-杨紫嫣Cynthia---印花衬衫与黑色皮裙系列-60P-
-爱蜜社_-Vol.509-Jermy---少女的娇媚性感动人-37P-
-爱蜜社_-Vol.510-九月生_---轻薄精致的睡衣包裹-44P-
-爱蜜社_-Vol.511-Angela小热巴---典雅吊裙勾勒出完美的曲线-55P-
-爱蜜社_-Vol.512-方子萱---洁白飒爽的职业西服-60P-
-爱蜜社_-Vol.513-Lavinia肉肉---丝袜美腿特辑写真-49P-
-爱蜜社_-Vol.514-Jermy---黑色衬衫与极致丝袜美腿-47P-
-爱蜜社_-Vol.515-九月生_---旗袍与极致魅惑的黑丝-41P-
-爱蜜社_-Vol.516-Evon陈赞之---凹凸有致的丰满身材-53P-
-爱蜜社_-Vol.517-杨紫嫣Cynthia---华丽典雅的礼裙-39P-
-爱蜜社_-Vol.518-Angela小热巴---丝足美腿灵动曼妙-45P-
-爱蜜社_-Vol.519--Lavinia肉肉--Jermy---OL短裙与现代性感丝袜系列-55P-
-爱蜜社_-Vol.520-萌白酱---经典的职场秘书OL制服系列-50P-
-爱蜜社_-Vol.521-Angela小热巴---都市风格的职业OL系列-54P-
-爱蜜社_-Vol.522-Sia---自有一股秀雅的气质-77P-
-爱蜜社_-Vol.523-九月生_---室外泳池系列-64P-
-爱蜜社_-Vol.524-萌白酱---深受男士喜欢的空姐制服系列-60P-
-爱蜜社_-Vol.525-气质美女-Vanessa---丝袜美腿写真-56P-
-爱蜜社_-Vol.526-Angela小热巴---靓丽车拍与极致朦胧美腿系列-42P-
-爱蜜社_-Vol.527-萌白酱---高挑婀娜多姿下的丝足美腿-61P-
-爱蜜社_-Vol.528-Vanessa---经典典雅职场制服-71P-
-爱蜜社_-Vol.529-Lavinia肉肉---朦胧丝袜美腿修长灵动诱人-53P-
-爱蜜社_-Vol.530-杨紫嫣Cynthia---深绿色华丽典雅的吊裙-56P-
-爱蜜社_-Vol.531-Angela小热巴---米色内衣与朦胧丝袜玉足-46P-
-爱蜜社_-Vol.532-Vanessa---浅蓝色的典雅职业装丝袜美腿-48P-
-爱蜜社_-Vol.533-Lavinia肉肉---娇艳俏丽的面容与妩媚得体姿态-49P-
-爱蜜社_-Vol.534-杨紫嫣Cynthia---古典韵味旗袍与现代性感丝袜系列-64P-
-爱蜜社_-Vol.535-Lavinia肉肉---洁白典雅的服饰轻裹娇躯-54P-
-爱蜜社_-Vol.536-Jermy---魅惑黑丝美腿-34P-
-爱蜜社_-Vol.537-Lavinia肉肉---光润玉颜的装束，妩媚得体姿态-48P-
-爱蜜社_-Vol.538-Vanessa---性感OL写真-60P-
-爱蜜社_-Vol.539-九月生_---经典的制服系列-44P-
-爱蜜社_Winki丝姬《街拍丝袜美腿-车内丝袜诱惑》-Vol.094-写真集-34P-
-爱蜜社_七七、张咏儿rainny---合集-Vol.002-写真集-64P-
-爱蜜社_乐薇Vivi《黑丝美腿诱惑》-Vol.293-写真集-44P-
-爱蜜社_九尾ivy-《空乘制服丝袜美腿诱惑》-Vol.276-写真集-47P-
-爱蜜社_九月生_《白衬衫短裙秘书职业装丝袜美足系列》-Vol.435-写真集-60P-
-爱蜜社_于姬Una《白皙的皮肤，呼之欲出的爆乳》-Vol.061-写真集-51P-
-爱蜜社_刘奕宁Lynn《巴厘岛旅拍》2套泳装-Vol.112-写真集-49P-
-爱蜜社_刘奕宁Lynn《美腿短裙-性感内衣》-Vol.088-写真集-43P-
-爱蜜社_夏茉GIGI《家居小清凉-性感猫女装》-Vol.024-写真集-55P-
-爱蜜社_夏茉GIGI《小清新街拍》-Vol.053-写真集-50P-
-爱蜜社_夏茉GIGI《清凉比基尼街拍》-Vol.020-写真集-54P-
-爱蜜社_夏茉GIGI《牛仔热裤街拍》-Vol.049-写真集-47P-
-爱蜜社_女神-Lavinia肉肉《丝袜美腿特辑》-Vol.401-写真集-42P-
-爱蜜社_妍妍Yanyan《同程旅拍》-Vol.008-写真集-76P-
-爱蜜社_孙梦瑶V《丝袜美腿控》-Vol.232-写真集-33P-
-爱蜜社_小九月《软妹控-丝袜控》-Vol.399-写真集-42P-
-爱蜜社_小宝酱《室外车拍白色丝袜诱惑》-Vol.056-写真集-44P-
-爱蜜社_小宝酱《牛仔美少女》-Vol.026-写真集-50P-
-爱蜜社_小狐狸Sica《黑色私房性感美腿》-VOL.194-写真集-51P-
-爱蜜社_張詠兒_rainny《内衣私房》-Vol.022-写真集-55P-
-爱蜜社_心妍小公主《极致诱惑黑丝旗袍内衣系列》-Vol.379-写真集-32P-
-爱蜜社_心妍小公主《魅惑黑丝内衣与极致丝袜系列》-Vol.421-写真集-28P-
-爱蜜社_戴小唯《比基尼-死库水系列》-Vol.085-写真集-49P-
-爱蜜社_李李七七喜喜《3套性感服饰》-Vol.081-写真集-48P-
-爱蜜社_李李七七喜喜《丝袜美腿-大尺度真空锁链》-VOL.203-写真集-48P-
-爱蜜社_李玉洁Daisy《武汉大别山薄刀峰旅拍》-Vol.040-写真集-57P-
-爱蜜社_杨晨晨sugar《丝袜美腿诱惑-完善礼物》-Vol.181-写真集-45P-
-爱蜜社_杨晨晨sugar《甜美的护士女神》-Vol.283-写真集-37P-
-爱蜜社_杨紫嫣Cynthia《正装之下的丝袜美腿》-Vol.428-写真集-40P-
-爱蜜社_极品长腿美女-刘奕宁Lynn-Vol.080-写真集-49P-
-爱蜜社_果儿Victoria《透视肚兜、黑丝捆绑极致诱惑销魂》-Vol.355-写真集-64P-
-爱蜜社_樂樂Mango《3套透视装》-Vol.077-写真集-45P-
-爱蜜社_樂樂Mango《丝袜美腿系列》-VOL.196-写真集-45P-
-爱蜜社_樂樂Mango《海边比基尼系列》-Vol.014-写真集-50P-
-爱蜜社_樂樂Mango《白衬衣-紫色丝袜系列》-Vol.046-写真集-57P-
-爱蜜社_模特Vissa《独特的熟女气质》-Vol.257-写真集-62P-
-爱蜜社_模特-小热巴《最爱的性感腰线》-Vol.241-写真集-51P-
-爱蜜社_沈佳熹《原生态人体诱惑》-Vol.114-写真集-20P-
-爱蜜社_沈佳熹《巴厘岛旅拍》死库水-比基尼-Vol.109-写真集-51P-
-爱蜜社_沈蜜桃miko《日本旅拍》-Vol.244-写真集-42P-
-爱蜜社_王曼妮好Q《性感睡衣-白衬衫系列》-Vol.137-写真集-45P-
-爱蜜社_王曼妮好Q-《街拍-泳池旁小清新系列》-Vol.141-写真集-40P-
-爱蜜社_王曼妮、谭小雅《古典旗袍系列》-Vol.070-写真集-60P-
-爱蜜社_王沫儿SaSa《丝袜美腿-傲人巨乳》-Vol.128-写真集-50P-
-爱蜜社_王沫儿SaSa《巨乳诱惑与别样胶带捆绑》-Vol.130-写真集-45P-
-爱蜜社_程小烦不烦《2套美胸泳装》-Vol.103-写真集-55P-
-爱蜜社_程小烦不烦《3套泳装》-Vol.067-写真集-54P-
-爱蜜社_程小烦不烦《性感比基尼和白衬衫-短裙》-Vol.089-写真集-47P-
-爱蜜社_程小烦不烦《皮卡丘-比基尼》-Vol.057-写真集-53P-
-爱蜜社_程小烦不烦《粉色装扮-白色裙子外拍》-Vol.086-写真集-47P-
-爱蜜社_美蒂Melody-Vol.280-写真集-39P-
-爱蜜社_艾诗iccey《淑女的诱惑》-Vol.036-写真集-49P-
-爱蜜社_芝芝Booty《丁字裤美臀-粉嫩和服》-VOL.155-写真集-50P-
-爱蜜社_芝芝Booty《日系和服-迷你比基尼》-VOL.170-写真集-37P-
-爱蜜社_莉莉Lily《兔女郎、空姐制服-死库水湿身系列》-Vol.082-写真集-49P-
-爱蜜社_萌琪琪Irene《黑丝美腿-街拍OL萌软妹》-VOL.199-写真集-64P-
-爱蜜社_许诺Sabrina《北海道旅拍》-Vol.213-写真集-34P-
-爱蜜社_许诺Sabrina《美得象一首抒情诗》-Vol.262-写真集-42P-
-爱蜜社_赵小米Kitty、sugar小甜心CC《闺蜜》-Vol.133-写真集-31P-
-爱蜜社_赵小米Kitty《比基尼-愤怒的宠物精灵》-Vol.110-写真集-54P-
-爱蜜社_陈思琪Art《凹凸别致的惹火身段》-Vol.345-写真集-37P-
-爱蜜社_陈思琪Art《白色连衣裙和内衣系列》-Vol.052-写真集-50P-
-爱蜜社_陈思琪Art《紧身裤-睡裙》-Vol.035-写真集-57P-
-爱蜜社_陈思琪Art《长腿OL-内衣》-Vol.042-写真集-59P-
-爱蜜社_陈良玲Carry《韵味黑色旗袍与魅惑黑丝》-Vol.358-写真集-42P-
-爱蜜社_雪儿Cier《肉色丝袜与美臀极致销魂》-Vol.325-写真集-35P-
-爱蜜社_香港模特Michi_Ho-《小清新》-Vol.045-写真集-48P-
-爱蜜社_高挑美女SISY思《4套清新服饰》-Vol.041-写真集-61P-
-爱蜜社_黄楽然《美胸翘臀丝袜美腿诱惑》-Vol.303-写真集-47P-
-爱蜜社_黄歆苑《性感内衣-学生装》-Vol.074-写真集-50P-
-爱蜜社_黄歆苑《泳装-蕾丝旗袍-长裙美腿》--Vol.075-写真集-48P-
-爱蜜社_黄歆苑《猫耳-性感内衣-黑丝SM》-Vol.076-写真集-68P-
+瑞丝馆_Breeze奕《性感黑丝内衣下曼妙多姿的娇躯》-Vol.059-写真集-41P-
+瑞丝馆_Buylu《身娇体柔的软妹子》-Vol.089-写真集-48P-
+瑞丝馆_Daisy琳琳《清新甜美的妹子》-Vol.062-写真集-42P-
+瑞丝馆_Daisy琳琳《清新甜美的妹子》-Vol.082-写真集-48P-
+瑞丝馆_Egg_尤妮丝《大胆私房诱惑》-Vol.053-写真集-35P-
+瑞丝馆_Eileen爱琳《身姿娇小但玲珑有致的软妹子》-Vol.072-写真集-40P-
+瑞丝馆_Eileen爱琳《身姿娇小但玲珑有致的软妹子》-Vol.075-写真集-44P-
+瑞丝馆_K8傲娇萌萌Vivian《不同颜色的丝袜妩媚诱惑》-VOL.011-写真集-52P-
+瑞丝馆_K8傲娇萌萌Vivian《丝袜主题写真》-VOL.017-写真集-63P-
+瑞丝馆_M梦baby《泡泡浴》-Vol.052-写真集-38P-
+瑞丝馆_M梦baby《魅惑白衬衫》-Vol.051-写真集-39P-
+瑞丝馆_M梦baby《魅惑私房》-Vol.048-写真集-40P-
+瑞丝馆_Tobey轩轩酱《娇小可人的软妹子》-Vol.078-写真集-40P-
+瑞丝馆_人间不值得lily《身娇体柔的娇滴滴妹子》-Vol.087-写真集-42P-
+瑞丝馆_凯竹BuiBui《长裙街拍-蕾丝内衣》-Vol.013-写真集-47P-
+瑞丝馆_叶佳颐《美腿-浴室系列》-Vol.030-写真集-44P-
+瑞丝馆_周于希dummy《黑丝美腿丁字裤-浴室诱惑》-Vol.031-写真集-37P-
+瑞丝馆_周于希和闺蜜《私房缠绵美臀诱惑》--VOL.014-写真集-30P-
+瑞丝馆_圆圆Daphne《娇柔清纯可人的软妹子》-Vol.067-写真集-42P-
+瑞丝馆_圆圆Daphne《娇柔清纯可人的软妹子》-Vol.069-写真集-50P-
+瑞丝馆_土肥圆矮挫穷《5套情趣服饰主题》-Vol.042-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《5套服装主题》-Vol.046-写真集-44P-
+瑞丝馆_土肥圆矮挫穷《制服主题写真》-Vol.039-写真集-45P-
+瑞丝馆_土肥圆矮挫穷-周妍希《花样万种风情诱惑》-Vol.040-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《女仆主题》-Vol.028-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《女仆-开叉礼服-室外内衣》-Vol.029-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《暖光下拍摄的女神》-VOL.026-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《服装秀系列！》-Vol.038-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《浴巾裹胸系列》-Vol.015-写真集-20P-
+瑞丝馆_土肥圆矮挫穷《百变情趣内衣》-VOL.034-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《真空毛衣内衣》-Vol.032-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《美腿与丝袜的致命诱惑》-VOL.008-写真集-43P-
+瑞丝馆_土肥圆矮挫穷《诱人黑丝与妩媚小吊带系列》-VOL.033-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《魅惑丝袜》-VOL.036-写真集-45P-
+瑞丝馆_土肥圆矮挫穷《黑丝内衣诱惑》-VOL.035-写真集-45P-
+瑞丝馆_夏笑笑Summer《少女动心诱惑》-Vol.043-写真集-41P-
+瑞丝馆_夏笑笑Summer《柔美如玉的少女动心诱惑》-VOL.027-写真集-47P-
+瑞丝馆_女神-周妍希《服装主题写真》-Vol.041-写真集-45P-
+瑞丝馆_奶瓶土肥圆《丝袜美腿控的福利》-Vol.054-写真集-45P-
+瑞丝馆_奶瓶土肥圆矮挫丑黑穷《美妙妖娆的曲线》-Vol.055-写真集-51P-
+瑞丝馆_小葡萄miki《唯美不失性感》-Vol.085-写真集-43P-
+瑞丝馆_小葡萄miki《梦幻光线与诱人蕾丝的巧妙结合》-Vol.079-写真集-42P-
+瑞丝馆_巨乳女神王婉悠Queen-VOL.012-写真集-47P-
+瑞丝馆_性感女神-Egg-尤妮丝《诱人的邀请》-Vol.047-写真集-23P-
+瑞丝馆_性感女神-萌琪琪Irene-VOL.016-写真集-50P-
+瑞丝馆_慕西Mona《富含御姐气息的美女》-Vol.065-写真集-43P-
+瑞丝馆_慕西Mona《真空薄纱系列》-Vol.071-写真集-44P-
+瑞丝馆_新人模特画画sugar-VOL.004-写真集-50P-
+瑞丝馆_木木夕Mmx《各式性感女仆厨娘制服系列》-Vol.064-写真集-42P
+瑞丝馆_木木夕Mmx《唯美性感内衣系列》-Vol.081-写真集-46P-
+瑞丝馆_木木夕Mmx《娇柔可人温柔乖巧的清纯妹子》-Vol.063-写真集-40P-
+瑞丝馆_木木夕Mmx《性感内衣系列》-Vol.070-写真集-46P-
+瑞丝馆_木木夕Mmx《花样百变的主题服饰系列》-Vol.083-写真集-53P-
+瑞丝馆_李可可《蕾丝镂空半透内衣》-Vol.056-写真集-48P-
+瑞丝馆_李可可《蕾丝镂空半透内衣装扮》-Vol.058-写真集-50P-
+瑞丝馆_楚楚baby《材高挑多姿颜值清秀精致的美女》-Vol.057-写真集-50P-
+瑞丝馆_模特M梦baby《柔情私房》-Vol.045-写真集-40P-
+瑞丝馆_模特M梦baby《黑丝吊袜》-Vol.050-写真集-40P-
+瑞丝馆_浪漫果味《新人模特首套》-Vol.073-写真集-35P-
+瑞丝馆_深紫Julie《私房内的清甜可人草帽女孩系列》-Vol.068-写真集-47P-
+瑞丝馆_深紫Julie《首套写真》-Vol.066-写真集-39P-
+瑞丝馆_清纯妹子-玟姊BABY首套私房-VOL.007-写真集-50P-
+瑞丝馆_清纯甜美妹子-米粒sweet-Vol.037-写真集-40P-
+瑞丝馆_潘琳琳ber《巨乳粉红女郎》-Vol.076-写真集-37P-
+瑞丝馆_潘琳琳ber《巨乳肥臀尤物》-Vol.086-写真集-35P-
+瑞丝馆_潘琳琳ber《巨乳肥臀熟女气息十足》-Vol.074-写真集-30P-
+瑞丝馆_潘琳琳ber《巨乳肥臀熟女气息十足》-Vol.077-写真集-35P-
+瑞丝馆_白沫《身材苗条美腿修长诱人》-Vol.060-写真集-37P-
+瑞丝馆_筱慧-Vol.088-写真集-43P-
+瑞丝馆_筱慧《白衬衫丝袜牛仔裤的精彩美臀诱惑》-Vol.084-写真集-46P-
+瑞丝馆_美女主播-周于希dummy-VOL.003-写真集-45P-
+瑞丝馆_美女主播-孙梦瑶V-VOL.018-写真集-45P-
+瑞丝馆_萌琪琪Irene《丝袜诱惑系列》-VOL.005-写真集-47P-
+瑞丝馆_萌琪琪Irene《兔女郎与性感猫耳》-VOL.009-写真集-50P-
+瑞丝馆_萌琪琪Irene《皮鞭、黑丝、兔女郎、空姐等服饰》-VOL.001-写真集-50P-
+瑞丝馆_萌琪琪Irene《花样性感情趣内衣》-VOL.021-写真集-50P-
+瑞丝馆_萌琪琪Irene《透视性感内衣-情趣丝袜》-VOL.025-写真集-50P-
+瑞丝馆_萌琪琪Irene《透视旗袍、黑丝美腿》-VOL.006-写真集-50P-
+瑞丝馆_豆浆妹《颜值跟身材都很不错的妹子》-Vol.061-写真集-65P-
+瑞丝馆_雪瑞Lisa《不同颜色的丝袜与眼镜OL》-VOL.024-写真集-46P-
+瑞丝馆_雪瑞Lisa《丝袜主题写真》-VOL.020-写真集-51P-
+瑞丝馆_雪瑞Lisa《低胸睡衣、女仆装、学生装、蕾丝袜》-VOL.022-写真集-54P-
+瑞丝馆_雪瑞Lisa《肚兜-黑丝诱惑》-VOL.010-写真集-52P-
+瑞丝馆_雪瑞Lisa《诱人身材与丝袜诱惑》-VOL.019-写真集-52P-
+瑞丝馆_顾北北《少女的芳香》-VOL.023-写真集-40P-
+瑞丝馆_风衣月Lucy《肌肤跟身材都不错的妹子》-Vol.080-写真集-46P-
+瑞丝馆_麦提娜Mitina《呼之欲出的美胸》-VOL.002-写真集-50P-
 EOF;
 
 // 重命名文件夹用tujigu的名字，copy desc 和 tags，重命名thumbnail
@@ -197,6 +122,8 @@ if(empty($folder_full_path)) {
     exit;
 }
 
+$dryrun = 0;
+
 $scan = scandir($folder_full_path);
 foreach($scan as $folder_name) {
    if (is_dir("$folder_full_path/$folder_name") && $folder_name!='.' && $folder_name!='..') {
@@ -216,32 +143,38 @@ foreach($scan as $folder_name) {
         if(!empty($origin)) {
             // copy desc.txt and tags.txt from origin folder
             echo "copy " . $origin['full_path'] . "/desc.txt " . $the_full_path . "/\n";
-            copy($origin['full_path'] . "/desc.txt", $the_full_path . "/desc.txt");
+            if(!$dryrun) copy($origin['full_path'] . "/desc.txt", $the_full_path . "/desc.txt");
             
             // copy($origin['full_path'] . "/tags.txt", $the_full_path . "/tags.txt");
 
             // Gen new tags
             $tags_str = file_get_contents($origin['full_path'] . "/tags.txt");
             $tags = explode(',', $tags_str);
-            if(! in_array('IMiss', $tags) && ! in_array('imiss', $tags)) {
-                $tags[] = 'IMiss';
+            if(! in_array('RuiSG', $tags) && ! in_array('ruisg', $tags)) {
+                $tags[] = 'RuiSG';
             }
-            if(! in_array( '爱蜜社', $tags ) ) {
-                $tags[] = '爱蜜社';
+            if(! in_array( '瑞丝馆', $tags ) ) {
+                $tags[] = '瑞丝馆';
+            }
+            if(! in_array('MiCat', $tags) && ! in_array('micat', $tags)) {
+                $tags[] = 'MiCat';
+            }
+            if(! in_array( '猫萌榜', $tags ) ) {
+                $tags[] = '猫萌榜';
             }
             if(! in_array( $origin['model'], $tags ) ) {
                 $tags[] = $origin['model'];
             }
             $new_tags_str = implode(',', $tags);
-            $new_tags_str = str_replace('imiss', 'IMiss', $new_tags_str);
+            $new_tags_str = str_replace('ruisg', 'RuiSG', $new_tags_str);
             echo "---- New tags: $new_tags_str \n";
             // create tags.txt
             echo "file_put_contents($the_full_path/tags.txt, $new_tags_str) \n";
-            file_put_contents($the_full_path.'/tags.txt', $new_tags_str);
+            if(!$dryrun) file_put_contents($the_full_path.'/tags.txt', $new_tags_str);
 
             // create a new empty dl.txt in current folder
             echo "file_put_contents($the_full_path/dl.txt, '') \n";
-            file_put_contents($the_full_path.'/dl.txt', '');
+            if(!$dryrun) file_put_contents($the_full_path.'/dl.txt', '');
 
             // If no thumbnail.jpg, rename last image to thumbnail.jpg
             $scan2 = scandir($the_full_path);
@@ -252,18 +185,18 @@ foreach($scan as $folder_name) {
                     $images[] = $file;
                 }
             }
-            if((int)$vol > 34 && (int)$vol < 282) {
+            // if((int)$vol > 34 && (int)$vol < 282) {
                 $thumb = $images[0];
-            }
-            else {
-                $thumb = $images[count($images) - 1];
-            }
+            // }
+            // else {
+                // $thumb = $images[count($images) - 1];
+            // }
             echo 'rename('. $the_full_path . '/' . $thumb . ', ' . $the_full_path . '/thumbnail.jpg' .  "\n\n";
-            rename($the_full_path . '/' . $thumb , $the_full_path . '/thumbnail.jpg');    
+            if(!$dryrun) rename($the_full_path . '/' . $thumb , $the_full_path . '/thumbnail.jpg');    
 
             // rename current folder to the origin folder name
-            echo "rename $the_full_path to $folder_full_path/IMiss爱蜜社-". $vol. '-'. $origin['main'] . "\n\n";
-            rename($the_full_path, $folder_full_path . '/IMiss爱蜜社-' . $vol. '-'. $origin['main']);
+            echo "rename $the_full_path to $folder_full_path/MiCat猫萌榜-". $vol. '-'. $origin['main'] . "\n\n";
+            if(!$dryrun) rename($the_full_path, $folder_full_path . '/MiCat猫萌榜-' . $vol. '-'. $origin['main']);
         }
    }
 }
@@ -271,12 +204,14 @@ foreach($scan as $folder_name) {
 
 
 function getOriginFolderByVol($vol_num) {
-    global $dirs;
     // $vol_num = '第' . (int)$vol_num . '期';
     // echo "vol_num = $vol_num \n";
-    $origin_folders = '/mnt/nas/jw-photos/tujigu/爱蜜社';
+    
+    $origin_folders = '/mnt/nas/jw-photos/tujigu/瑞丝馆';
     // $scan = scandir($origin_folders);
+    global $dirs;
     $scan = explode("\r\n", $dirs);
+
     $folder_info = [];
     
     foreach($scan as $folder_name) {      
@@ -294,7 +229,7 @@ function getOriginFolderByVol($vol_num) {
                 $folder_info['name'] = $l_folder_name;
                 $folder_info['full_path'] = $origin_folders  . '/' . $folder_name;
 
-                $tt = find_between($l_folder_name, '爱蜜社_', '-');
+                $tt = find_between($l_folder_name, '瑞丝馆_', '-');
                 if(str_starts_with($tt[0], 'vol.')) {
                     $ss = explode('-', $l_folder_name);
                     $folder_info['model'] = $ss[1];
@@ -306,7 +241,7 @@ function getOriginFolderByVol($vol_num) {
                 }
                 else {
                     $folder_info['model'] = $tt[0];
-                    $yy = find_between($l_folder_name, '爱蜜社_', '-vol');
+                    $yy = find_between($l_folder_name, '瑞丝馆_', '-vol');
                     $main = trim($yy[0], '-');
                     $folder_info['main'] = $main;
                 }
